relocate fetch init stake and hotshot config
