+++ contracts/rust/deployer/Cargo.toml
@@ -10,7 +10,11 @@ alloy = { workspace = true }
 anyhow = { workspace = true }
 clap = { workspace = true }
 derive_more = { workspace = true }
+espresso-types = { path = "../../../types" }
 hotshot-contract-adapter = { workspace = true }
 hotshot-types = { workspace = true }
+surf-disco = { workspace = true }
+tide-disco = { workspace = true }
 tokio = { workspace = true }
 tracing = "0.1.37"
+vbs = { workspace = true }
