+++ hotshot-state-prover/src/bin/gen-demo-genesis.rs
@@ -1,8 +1,8 @@
 use alloy::{hex::ToHexExt, sol_types::SolValue};
 use clap::Parser;
+use espresso_contract_deployer::network_config::light_client_genesis;
 use hotshot_contract_adapter::sol_types::{LightClientStateSol, StakeTableStateSol};
 use hotshot_stake_table::config::STAKE_TABLE_CAPACITY;
-use hotshot_state_prover::service::light_client_genesis;
 use url::Url;
 
 #[derive(Parser)]
